# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MemMachine API",
    "version": "1.0.0",
    "description": "API endpoints for interacting with the Profile Memory and Episodic Memory components of the MemMachine memory system."
  },
  "servers": [
    {
      "url": "https://localhost:8000/v1"
    }
  ],
...
  "paths": {
    "/memories": {
      "post": {
        "tags": [
          "RESTful APIs"
        ],
        "summary": "/memories",
        "description": "This endpoint first retrieves the appropriate episodic memory instance based on the session context (group, agent, user, session IDs). It then adds the episode to the episodic memory. If successful, it also passes the message to the profile memory for ingestion.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewEpisode"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Memory successfully added."
          },
          "400": {
            "description": "Invalid producer or produced_for ID."
          },
          "404": {
            "description": "No matching episodic memory instance found."
          }
        }
      },
      "delete": {
        "tags": [
          "RESTful APIs"
        ],
        "summary": "/memories",
        "description": "Deletes all data for a particular session.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteDataRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session data successfully deleted."
          },
          "404": {
            "description": "No matching episodic memory instance found."
          }
        }
      }
    },
    "/memories/search": {
      "post": {
        "tags": [
          "RESTful APIs"
        ],
        "summary": "/memories/search",
        "description": "Retrieves the relevant episodic memory instance and then performs concurrent searches in both the episodic memory and the profile memory. The results are combined into a single response object.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Search results returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResult"
                }
              }
            }
          },
          "404": {
            "description": "No matching episodic memory instance found."
          }
        }
      }
    },
...
    "/users/{user_id}/sessions": {
      "get": {
        "tags": [
          "RESTful APIs"
        ],
        "summary": "/users/{user_id}/sessions",
        "description": "Get all sessions for a particular user",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of sessions for the specified user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllSessionsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/groups/{group_id}/sessions": {
      "get": {
        "tags": [
          "RESTful APIs"
        ],
        "summary": "/groups/{group_id}/sessions",
        "description": "Get all sessions for a particular group",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of sessions for the specified group.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllSessionsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/agents/{agent_id}/sessions": {
      "get": {
        "tags": [
          "RESTful APIs"
        ],
        "summary": "/agents/{agent_id}/sessions",
        "description": "Get all sessions for a particular agent",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of sessions for the specified agent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllSessionsResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/mcp/users/{user_id}/sessions": {
      "get": {
        "tags": [
          "MCP Server APIs"
        ],
        "summary": "users/{user_id}/sessions",
        "description": "Retrieves all sessions for a specific user. This endpoint is for internal use by LLMs via FastMCP.",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of sessions for the user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllSessionsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/mcp/groups/{group_id}/sessions": {
      "get": {
        "tags": [
          "MCP Server APIs"
        ],
        "summary": "groups/{group_id}/sessions",
        "description": "Retrieves all sessions for a specific group. This endpoint is for internal use by LLMs via FastMCP.",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of sessions for the group.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllSessionsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/mcp/agents/{agent_id}/sessions": {
      "get": {
        "tags": [
          "MCP Server APIs"
        ],
        "summary": "agents/{agent_id}/sessions",
        "description": "Retrieves all sessions for a specific agent. This endpoint is for internal use by LLMs via FastMCP.",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of sessions for the agent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllSessionsResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "SessionData": {
        "type": "object",
        "properties": {
          "group_id": {
            "type": "string",
            "nullable": true
          },
          "agent_id": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "user_id": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "session_id": {
            "type": "string"
          }
        }
      },
      "NewEpisode": {
        "type": "object",
        "properties": {
          "session": {
            "$ref": "#/components/schemas/SessionData"
          },
          "producer": {
            "type": "string"
          },
          "produced_for": {
            "type": "string"
          },
          "episode_content": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "number"
                }
              }
            ]
          },
          "episode_type": {
            "type": "string"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true,
            "nullable": true
          }
        }
      },
      "SearchQuery": {
        "type": "object",
        "properties": {
          "session": {
            "$ref": "#/components/schemas/SessionData"
          },
          "query": {
            "type": "string"
          },
          "filter": {
            "type": "object",
            "additionalProperties": true,
            "nullable": true
          },
          "limit": {
            "type": "integer",
            "nullable": true
          }
        }
      },
      "SearchResult": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "default": 0
          },
          "content": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "DeleteDataRequest": {
        "type": "object",
        "properties": {
          "session": {
            "$ref": "#/components/schemas/SessionData"
          }
        }
      },
      "MemorySession": {
        "type": "object",
        "properties": {
          "user_ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "session_id": {
            "type": "string"
          },
          "group_id": {
            "type": "string",
            "nullable": true
          },
          "agent_ids": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        }
      },
      "AllSessionsResponse": {
        "type": "object",
        "properties": {
          "sessions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemorySession"
            }
          }
        }
      }
    }
  }
}